refactor 2

+++ packages/create-waku/src/index.ts
@@ -1,13 +1,13 @@
 import { existsSync, readdirSync } from 'node:fs';
 import fsPromises from 'node:fs/promises';
 import path from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { parseArgs } from 'node:util';
 import { default as prompts } from 'prompts';
 import { red, green, bold } from 'kolorist';
 import fse from 'fs-extra/esm';
 import checkForUpdate from 'update-check';
 import { createRequire } from 'node:module';
+import { installTemplate } from './helpers/install-template.js';
 import {
   parseExampleOption,
   downloadAndExtract,
@@ -29,122 +29,29 @@ const { values } = parseArgs({
   },
 });
 
-function isValidPackageName(projectName: string) {
-  return /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/.test(
-    projectName,
-  );
-}
-
-function toValidPackageName(projectName: string) {
-  return projectName
-    .trim()
-    .toLowerCase()
-    .replace(/\s+/g, '-')
-    .replace(/^[._]/, '')
-    .replace(/[^a-z0-9-~]+/g, '-');
-}
-
-// if the dir is empty or not exist
-function canSafelyOverwrite(dir: string) {
-  return !existsSync(dir) || readdirSync(dir).length === 0;
-}
-
-function displayUsage() {
-  const cmd = process.argv.slice(0, 2).join(' ');
-  console.log(`
-Usage: ${cmd} [options]
-
-Options:
-  --example             Specify an example use as a template
-  -h, --help            Display this help message
-`);
-}
-
-async function notifyUpdate() {
-  // keep original require to avoid
-  //  bundling the whole package.json by `@vercel/ncc`
-  const packageJson = createRequire(import.meta.url)('../package.json');
-  const result = await checkForUpdate(packageJson).catch(() => null);
-  if (result?.latest) {
-    console.log(`A new version of 'create-waku' is available!`);
-    console.log('You can update by running: ');
-    console.log();
-    console.log(`    npm i -g create-waku`);
-  }
-}
-
-async function installTemplate({
-  root,
-  packageName,
-}: {
-  root: string;
-  packageName: string;
-}) {
-  const pkg = {
-    name: packageName,
-    version: '0.0.0',
-  };
-
-  const templateRoot = path.join(
-    fileURLToPath(import.meta.url),
-    '../../template',
-  );
-  // maybe include `.DS_Store` on macOS
-  const CHOICES = (await fsPromises.readdir(templateRoot)).filter(
-    (dir) => !dir.startsWith('.'),
-  );
-
-  const templateDir = path.join(templateRoot, CHOICES[0]!);
-
-  // Read existing package.json from the root directory
-  const packageJsonPath = path.join(root, 'package.json');
-
-  // Read new package.json from the template directory
-  const newPackageJsonPath = path.join(templateDir, 'package.json');
-  const newPackageJson = JSON.parse(
-    await fsPromises.readFile(newPackageJsonPath, 'utf-8'),
-  );
-
-  fse.copySync(templateDir, root);
-
-  await fsPromises.writeFile(
-    packageJsonPath,
-    JSON.stringify(
-      {
-        ...newPackageJson,
-        ...pkg,
-      },
-      null,
-      2,
-    ),
-  );
-
-  if (existsSync(path.join(root, 'gitignore'))) {
-    await fsPromises.rename(
-      path.join(root, 'gitignore'),
-      path.join(root, '.gitignore'),
+async function doPrompts() {
+  const isValidPackageName = (projectName: string) =>
+    /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/.test(
+      projectName,
     );
-  }
-}
 
-async function init() {
-  if (values.help) {
-    displayUsage();
-    return;
-  }
+  const toValidPackageName = (projectName: string) =>
+    projectName
+      .trim()
+      .toLowerCase()
+      .replace(/\s+/g, '-')
+      .replace(/^[._]/, '')
+      .replace(/[^a-z0-9-~]+/g, '-');
 
-  const repoInfo = await parseExampleOption(values.example, DEFAULT_REF);
+  // if the dir is empty or not exist
+  const canSafelyOverwrite = (dir: string) =>
+    !existsSync(dir) || readdirSync(dir).length === 0;
 
-  let targetDir = '';
   const defaultProjectName = 'waku-project';
-
-  let result: {
-    packageName: string;
-    shouldOverwrite: string;
-  };
+  let targetDir = '';
 
   try {
-    result = (await prompts(
+    const result = await prompts(
       [
         {
           name: 'projectName',
@@ -182,18 +89,56 @@ async function init() {
           throw new Error(red('✖') + ' Operation cancelled');
         },
       },
-    )) as any; // FIXME no-any
-  } catch (cancelled) {
-    if (cancelled instanceof Error) {
-      console.log(cancelled.message);
+    );
+    return {
+      ...result,
+      packageName: result.packageName ?? toValidPackageName(targetDir),
+      targetDir,
+    };
+  } catch (err) {
+    if (err instanceof Error) {
+      console.log(err.message);
     }
     process.exit(1);
   }
+}
+
+function displayUsage() {
+  const cmd = process.argv.slice(0, 2).join(' ');
+  console.log(`
+Usage: ${cmd} [options]
+
+Options:
+  --example             Specify an example use as a template
+  -h, --help            Display this help message
+`);
+}
+
+async function notifyUpdate() {
+  // keep original require to avoid
+  //  bundling the whole package.json by `@vercel/ncc`
+  const packageJson = createRequire(import.meta.url)('../package.json');
+  const result = await checkForUpdate(packageJson).catch(() => null);
+  if (result?.latest) {
+    console.log(`A new version of 'create-waku' is available!`);
+    console.log('You can update by running: ');
+    console.log();
+    console.log(`    npm i -g create-waku`);
+  }
+}
+
+async function init() {
+  if (values.help) {
+    displayUsage();
+    return;
+  }
+
+  const repoInfo = await parseExampleOption(values.example, DEFAULT_REF);
+
+  const { packageName, shouldOverwrite, targetDir } = await doPrompts();
+  const root = path.resolve(targetDir);
 
   console.log('Setting up project...');
-
-  const root = path.resolve(targetDir);
-  const { packageName, shouldOverwrite } = result;
 
   if (shouldOverwrite) {
     fse.emptyDirSync(root);
@@ -207,10 +152,7 @@ async function init() {
   } else {
     // If an example repository is not provided for cloning, proceed
     // by installing from a template.
-    await installTemplate({
-      root,
-      packageName: packageName ?? toValidPackageName(targetDir),
-    });
+    await installTemplate({ root, packageName });
   }
 
   const manager = process.env.npm_config_user_agent ?? '';
